Closes #2
Removed link to javascript in tag output

This was causing duplicate repos to appear as per #2

Added link to required javascript files in head

This allows the importing of the required javascript files. This also
relinquishes control from jquery back to various octopress plugins/asides
to use the `$` again.

Updated README and LICENCE

The README has updated (hopefully clear) instructions on how to install the
plugin.

The LICENCE contains the copyright notice of my added work.